requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.9.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
brotli==1.1.0
    # via whitenoise
celery==5.5.3
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2025.8.3
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
cron-descriptor==2.0.5
    # via django-celery-beat
dj-database-url==3.0.1
    # via -r requirements.in
dj-rest-auth==7.0.1
    # via -r requirements.in
django==5.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   dj-rest-auth
    #   django-allauth
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-redis
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   sentry-sdk
django-allauth==65.11.0
    # via -r requirements.in
django-celery-beat==2.8.1
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-storages==1.14.2
    # via -r requirements.in
django-timezone-field==7.1
    # via django-celery-beat
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
kombu==5.5.4
    # via celery
packaging==25.0
    # via
    #   gunicorn
    #   kombu
pillow==11.3.0
    # via -r requirements.in
prompt-toolkit==3.0.51
    # via click-repl
psutil==6.1.1
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.in
python-json-logger==4.0.0
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
redis==6.4.0
    # via
    #   -r requirements.in
    #   django-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via -r requirements.in
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
sentry-sdk[django]==2.17.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
structlog==25.4.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via cron-descriptor
tzdata==2025.2
    # via
    #   django
    #   django-celery-beat
    #   kombu
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise[brotli]==6.9.0
    # via -r requirements.in
